updates, mainly for gcc-2.95
git-svn-id: https://svn.r-project.org/R/trunk@5429 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/gnuwin32/extra.c
@@ -256,12 +256,12 @@ SEXP do_winver(SEXP call, SEXP op, SEXP args, SEXP env)
 	strcpy(isNT, "win32s");
 	break;	
     default:
-	sprintf(isNT, "ID=%d", verinfo.dwPlatformId);
+	sprintf(isNT, "ID=%d", (int)verinfo.dwPlatformId);
 	break;
     }
     
     sprintf(ver, "Windows %s %d.%d (build %d) %s", isNT,
-	    verinfo.dwMajorVersion, verinfo.dwMinorVersion,
+	    (int)verinfo.dwMajorVersion, (int)verinfo.dwMinorVersion,
 	    LOWORD(verinfo.dwBuildNumber), verinfo.szCSDVersion);
     
     PROTECT(ans = allocVector(STRSXP, 1));
